+++ Sources/NSManagedObjectContext+ManagedObject.swift
@@ -136,11 +136,33 @@ extension NSManagedObjectContext {
   /// - parameter queue: Optional operation queue on which to request the merge.
   /// - returns: An opaque observer object representing the merging block. Use
   ///   this to remove it.
-  func automaticallyMergesChanges(from context: NSManagedObjectContext, queue: OperationQueue? = nil) -> NSObjectProtocol {
+  public func automaticallyMergesChanges(from context: NSManagedObjectContext, queue: OperationQueue? = nil) -> NSObjectProtocol {
     let center = NotificationCenter.default
     let name = Notification.Name.NSManagedObjectContextDidSave
     return center.addObserver(forName: name, object: context, queue: queue) { [weak self] (notification) in
       self?.mergeChanges(fromContextDidSave: notification)
+    }
+  }
+
+  /// Sets up an observer block for context-will-save notifications. Updates
+  /// `createdAt` with the current date and time when the context sees a newly
+  /// inserted object. Updates `updatedAt` when the context sees an
+  /// update. Ignores deletions. Only stamps those objects that respond to
+  /// setters for the created- and updated-at dates.
+  public func automaticallyUpdatesTimestamps(queue: OperationQueue? = nil) -> NSObjectProtocol {
+    let center = NotificationCenter.default
+    let name = Notification.Name.NSManagedObjectContextWillSave
+    return center.addObserver(forName: name, object: self, queue: queue) { [weak self] (_) in
+      self?.insertedObjects.filter {
+        $0.responds(to: Selector(("setCreatedAt:")))
+      }.forEach {
+        $0.setValue(Date(), forKey: "createdAt")
+      }
+      self?.updatedObjects.filter {
+        $0.responds(to: Selector(("setUpdatedAt:")))
+      }.forEach {
+        $0.setValue(Date(), forKey: "updatedAt")
+      }
     }
   }